Add a more detailed example
I got about 2/3 of the way through this when I saw you had already done
this in your tests.

I was looking for something in ./examples that showed just about all of
the features of GraphQL, especially the Schema features. I couldn't find
it (until after I started) so I decided to implement the Star Wars
example from the Facebook Javascript GraphQL example.