[validation] Allow differing directives.
The currently supported directives: @skip and @include do not create an ambiguous or erroneous query when used on conflicting fields in a divergent fashion. In fact, this may be intended when many different conditions should result in the same outcome. For example:

```graphql
{
  field @include(if: $firstCondition)
  field @include(if: $secondCondition)
}
```

This example could be considered as the intent of fetching `field` given `$firstCondition || $secondCondition`. While this example is contrived, there are more complex examples where such fields are nested within fragments that this condition is reasonable.

Commit:
9b11df2efc66ad3c07e0d373a7e04a3ba5ee581a [9b11df2]
Parents:
f010e86c3d
Author:
Lee Byron <lee@leebyron.com>
Date:
13 November 2015 at 11:48:47 AM SGT